Wrote the TEST PLAN Created a plan for testing everything. Also moved the test database url to the environment, and told tox about it. Added one test for the models and moved all the fixtures to a conftest.py file.
