+++ pyramid_learning_journal/templates/create.jinja2
@@ -4,11 +4,11 @@
     <form method="POST">
         <div class="card mb-3">
             <div class="card-header bg-white">
-                <input type="text" name="title" placeholder="Title" class="form-control border-light bg-light">
+                <input type="text" name="title" placeholder="Title" class="form-control border-light bg-light" required>
             </div>
             <div class="card-body">
                 <div class="card-text detail">
-                    <textarea name="body" placeholder="Write a new entry..." class="form-control border-light bg-light"></textarea>
+                    <textarea name="body" placeholder="Write a new entry..." class="form-control border-light bg-light" required></textarea>
                 </div>
             </div>
         </div> <!-- end of card -->
