+++ pyramid_learning_journal/templates/edit.jinja2
@@ -4,12 +4,11 @@
     <form method='POST'>
         <div class="card mb-3">
             <div class="card-header bg-white">
-                <input type="text" name="title" placeholder="Title" value="{{ entry.title }}" class="form-control border-light bg-light">
+                <input type="text" name="title" placeholder="Title" value="{{ entry.title }}" class="form-control border-light bg-light" required>
             </div>
             <div class="card-body">
                 <div class="card-text detail">
-                    <textarea name="body" placeholder="Write a new entry..." class="form-control border-light bg-light">{{ entry.body }}
-                    </textarea>
+                    <textarea name="body" placeholder="Write a new entry..." class="form-control border-light bg-light" required>{{ entry.body }}</textarea>
                 </div>
             </div>
         </div> <!-- end of card -->
