+++ tox.ini
@@ -2,10 +2,9 @@
 envlist = py27, py36
 
 [testenv]
+passenv = TEST_DATABASE_URL
 commands = py.test --cov-report term-missing --cov=pyramid_learning_journal
 deps = 
     pytest
     pytest-cov
     WebTest
-    markdown
-    pytz
